STHG code/src/requirements.txt
absl-py==0.12.0
alembic==1.8.0
anykeystore==0.2
apex==0.9.10.dev0
apturl==0.5.2
async-generator==1.10
attrs==21.4.0
autopage==0.5.1
awscli==1.19.53
beautifulsoup4==4.9.3
bert==2.2.0
blinker==1.4
blis==0.7.4
boto3==1.17.51
botocore==1.20.51
Brlapi==0.7.0
bs4==0.0.1
cachetools==4.2.1
catalogue==2.0.2
cbor2==5.4.1
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.1.1
click==7.1.2
cliff==3.10.1
cmaes==0.8.2
cmd2==2.4.1
colorama==0.4.3
colorlog==6.6.0
command-not-found==0.3
conllu==4.4
cryptacular==1.5.5
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.5
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.12
dgl-cu102==0.8.2
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
docutils==0.15.2
en-core-sci-md==0.4.0
en-core-web-sm==3.0.0
entrypoints==0.3
erlastic==2.0.0
filelock==3.0.12
future==0.18.2
google-auth==1.29.0
google-auth-oauthlib==0.4.4
googledrivedownloader==0.4
gpg===1.13.1-unknown
greenlet==1.0.0
grpcio==1.37.0
h11==0.13.0
httplib2==0.14.0
hupper==1.10.2
idna==2.8
importlib-metadata==4.12.0
importlib-resources==5.8.0
isodate==0.6.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
luma.core==2.3.1
luma.oled==3.8.1
macaroonbakery==1.3.1
Mako==1.2.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.4.1
mpmath==1.2.1
murmurhash==1.0.5
netifaces==0.10.4
networkx==2.5.1
nltk==3.6.1
nmslib==2.1.1
numpy==1.20.2
oauthlib==3.1.0
olefile==0.46
optuna==2.10.1
outcome==1.2.0
packaging==21.3
pandas==1.2.4
PasteDeploy==2.1.1
pathlib==1.0.1
pathy==0.4.0
patsy==0.5.2
pbkdf2==1.3
pbr==5.9.0
pexpect==4.6.0
Pillow==7.0.0
plaster==1.0
plaster-pastedeploy==0.7
preshed==3.0.5
prettytable==2.1.0
protobuf==3.15.8
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pydantic==1.7.3
pyftdi==0.53.2
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyparsing==2.4.6
pyperclip==1.8.2
pyramid==2.0
pyramid-mailer==0.15.1
pyRFC3339==1.1
pysbd==0.3.4
pyserial==3.5
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-louvain==0.15
python3-openid==3.2.0
pytorch-lightning==0.9.0
pytorch-pretrained-bert==0.6.1
pytorch-transformers==1.2.0
pytz==2019.3
pyusb==1.1.0
pyxdg==0.26
PyYAML==5.3.1
rdflib==6.0.0
regex==2021.4.4
reportlab==3.5.34
repoze.sendmail==4.4.1
requests==2.28.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
RPi.GPIO==0.7.0
rsa==4.7.2
s3transfer==0.3.7
sacremoses==0.0.44
scikit-learn==0.24.1
scipy==1.6.2
scispacy==0.4.0
screen-resolution-extra==0.0.0
seaborn==0.11.1
SecretStorage==2.3.1
selenium==4.3.0
sentencepiece==0.1.95
simplejson==3.16.0
six==1.14.0
sklearn==0.0
smart-open==3.0.0
smbus2==0.4.1
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
spacy==3.0.5
spacy-legacy==3.0.2
spidev==3.5
SQLAlchemy==1.4.7
srsly==2.4.1
ssh-import-id==5.10
statsmodels==0.13.2
stevedore==3.5.0
sympy==1.8
systemd-python==234
tensorboard==2.2.0
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
thinc==8.0.2
threadpoolctl==2.1.0
tokenizers==0.10.2
torch==1.8.1
torch-geometric==1.7.2
torch-sparse==0.6.11
torch-struct==0.5
torchaudio==0.8.1
torchmetrics==0.9.3
torchvision==0.9.1
tqdm==4.60.0
transaction==3.0.1
transformers==4.5.1
translationstring==1.4
trio==0.21.0
trio-websocket==0.9.2
typer==0.3.2
typing==3.7.4.3
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufal.udpipe==1.2.0.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.9
velruse==1.1.1
venusian==3.0.0
wadllib==1.3.3
wasabi==0.8.2
wcwidth==0.2.5
WebOb==1.8.7
Werkzeug==1.0.1
wrapt==1.12.1
wsproto==1.1.0
WTForms==2.3.3
wtforms-recaptcha==0.3.2
xkit==0.0.0
zipp==3.8.0
zope.deprecation==4.4.0
zope.interface==5.3.0
zope.sqlalchemy==1.3
